requirements.txt
pandas
numpy
google-cloud-storage
scikit-learn
pyyaml
python-dotenv
ipykernel
seaborn
matplotlib
statsmodels